+++ R/supsc.R
@@ -18,7 +18,7 @@
 #' @param x A string to be converted to superscript.
 #' @return The superscript version of the string passed to the function,
 #' if one exists. Otherwise, a question mark will be returned.
-#' @seealso The \code{\link{subsc}} function to get subscripts.
+#' @family utf8
 #' @examples
 #' # Single letter
 #' paste0(supsc("a"), "Footnote")